Rename custom_lint_checks.py to pylint_extensions.py
